+++ src/src/MolecularMechanics/MolecularMechanicsDlg.c
@@ -363,7 +363,7 @@ static gboolean createPCGamessFiles(gint numberOfGeometries, ForceField** geomet
 			SAtomsProp prop = prop_atom_get(symbol);
 			fprintf(file,"%s %f %f %f %f\n", 
 				symbol,
-				(gfloat)prop.atomicNumber,
+				(gdouble)prop.atomicNumber,
 				geometries[i]->molecule.atoms[j].coordinates[0],
 				geometries[i]->molecule.atoms[j].coordinates[1],
 				geometries[i]->molecule.atoms[j].coordinates[2]
@@ -754,7 +754,7 @@ static gboolean runOnePCGamess(ForceField* geometry, gdouble* energy, gchar* fil
 		SAtomsProp prop = prop_atom_get(symbol);
 		fprintf(file,"%s %f %f %f %f\n", 
 			symbol,
-			(gfloat)prop.atomicNumber,
+			(gdouble)prop.atomicNumber,
 			geometry->molecule.atoms[j].coordinates[0],
 			geometry->molecule.atoms[j].coordinates[1],
 			geometry->molecule.atoms[j].coordinates[2]
