+++ src/src/OpenGL/AxisGL.h
@@ -2,7 +2,7 @@
 #ifndef __GABEDIT_AXISGL_H__
 #define __GABEDIT_AXISGL_H__
 
-void getAxisProperties(gboolean* show, gboolean* negative, gfloat origin[], gfloat* radius, gfloat* scal, gfloat xColor[], gfloat yColor[], gfloat zColor[]);
+void getAxisProperties(gboolean* show, gboolean* negative, gdouble origin[], gdouble* radius, gdouble* scal, gdouble xColor[], gdouble yColor[], gdouble zColor[]);
 void initAxis();
 void save_axis_properties();
 void read_axis_properties();
