+++ src/src/OpenGL/ContoursPov.h
@@ -2,7 +2,7 @@
 #ifndef __GABEDIT_CONTOURSPOV_H__
 #define __GABEDIT_CONTOURSPOV_H__
 
-gint addContoursPovRay(Grid* plansgrid,gint Ncontours,gfloat* values,gint i0,gint i1,gint numplan,gfloat gap);
+gint addContoursPovRay(Grid* plansgrid,gint Ncontours,gdouble* values,gint i0,gint i1,gint numplan,gdouble gap);
 void deleteContoursPovRayFile();
 
 #endif /* __GABEDIT_CONTOURSPOV_H__ */
